+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Add testing for Rails 7.1 by @MrSerth
 
+### Fixed
+
+- Invalid exception class for ActionDispatch::ExceptionWrapper by @jgraichen
+
 ## [1.16.0] - 2023-08-24
 
 ### Added